requirements.txt
Flask==3.1.0
Flask-Mail==0.10.0
Flask-WTF==1.2.1
Jinja2==3.1.6
Werkzeug==3.1.3
itsdangerous==2.2.0
blinker==1.9.0
click==8.1.8
MarkupSafe==3.0.2
python-dotenv==1.0.1